refactor: 重构项目结构，将工具函数迁移至composables目录
- 将useDataCode、useCheckDigit、useBle和useSerial从utils目录迁移至composables目录
- 删除旧的modal组件，使用新的Dialog组件替代
- 更新相关引用路径
- 更新auto-imports.d.ts以支持新的composables
