+++ src/components/Dialog/Setting.vue
@@ -1,7 +1,7 @@
 <script setup>
+import { useDataCode } from '@/composables/useDataCode/useDataCode'
 import { useSerialStore } from '@/store/useSerialStore'
 import { useSettingStore } from '@/store/useSettingStore'
-import { useDataCode } from '@/utils/useDataCode/useDataCode'
 import { dialogKeys, useDialog } from './composable'
 
 const { visible } = useDialog()
